+++ src/main/java/org/osgl/inject/util/ParameterizedTypeImpl.java
@@ -1,0 +1,72 @@
+package org.osgl.inject.util;
+
+/*-
+ * #%L
+ * OSGL Genie
+ * %%
+ * Copyright (C) 2016 - 2018 OSGL (Open Source General Library)
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+
+import org.osgl.$;
+
+import java.lang.reflect.ParameterizedType;
+import java.lang.reflect.Type;
+import java.util.Arrays;
+
+public class ParameterizedTypeImpl implements ParameterizedType {
+
+    private final Type[] actualTypeArguments;
+    private final Type   ownerType;
+    private final Type   rawType;
+
+    public ParameterizedTypeImpl(Type[] actualTypeArguments, Type ownerType, Type rawType){
+        this.actualTypeArguments = actualTypeArguments;
+        this.ownerType = ownerType;
+        this.rawType = rawType;
+    }
+
+    public Type[] getActualTypeArguments() {
+        return actualTypeArguments;
+    }
+
+    public Type getOwnerType() {
+        return ownerType;
+    }
+
+    public Type getRawType() {
+        return rawType;
+    }
+
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+
+        ParameterizedTypeImpl that = (ParameterizedTypeImpl) o;
+
+        // Probably incorrect - comparing Object[] arrays with Arrays.equals
+        if (!Arrays.equals(actualTypeArguments, that.actualTypeArguments)) return false;
+        if (ownerType != null ? !ownerType.equals(that.ownerType) : that.ownerType != null) return false;
+        return rawType != null ? rawType.equals(that.rawType) : that.rawType == null;
+
+    }
+
+    @Override
+    public int hashCode() {
+        return $.hc(actualTypeArguments, ownerType, rawType);
+    }
+}